refactor: post add dialog 분리
