+++ src/features/post/open-add-dialog/index.tsx
@@ -1,5 +1,5 @@
 import { useDialogStore } from "@/features/dialog/model/store"
-import { ADD_POST_DIALOG } from "../add-dialog"
+import { ADD_POST_DIALOG } from "../../../widgets/post/add-dialog"
 import { OpenPostAddDialogButton } from "@/entity/post/ui/OpenAddDialogButton"
 
 export const OpenPostAddDialog = () => {
